# docs/openapi.json
{
  "components": {
    "schemas": {
      "InspectionIn": {
        "additionalProperties": false,
        "description": "수입검사 한 건.",
        "properties": {
          "item_code": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Item Code",
            "type": "string"
          },
          "judged_by": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Judged By",
            "type": "string"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/MeasurementIn"
            },
            "title": "Measurements",
            "type": "array"
          },
          "nonconformity_code": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonconformity Code"
          },
          "quantity": {
            "minimum": 0.0,
            "title": "Quantity",
            "type": "number"
          },
          "received_date": {
            "format": "date",
            "title": "Received Date",
            "type": "string"
          },
          "special_acceptance": {
            "default": false,
            "title": "Special Acceptance",
            "type": "boolean"
          },
          "supplier_code": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Supplier Code",
            "type": "string"
          },
          "supplier_lot_number": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Supplier Lot Number",
            "type": "string"
          }
        },
        "required": [
          "item_code",
          "supplier_code",
          "supplier_lot_number",
          "quantity",
          "judged_by",
          "received_date"
        ],
        "title": "InspectionIn",
        "type": "object"
      },
      "InspectionOut": {
        "description": "판정이 끝난 뒤 남은 것.\n\n**로트 번호를 돌려준다.** 창고 라벨에 찍히는 번호이고, 부르는 쪽이 그것을\n다시 물어보게 하면 한 사실을 두 번 읽게 된다.",
        "properties": {
          "inspection_id": {
            "title": "Inspection Id",
            "type": "integer"
          },
          "ledger_entry_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ledger Entry Id"
          },
          "lot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Id"
          },
          "lot_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Number"
          },
          "nonconformity_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonconformity Code"
          },
          "result": {
            "enum": [
              "합격",
              "불합격",
              "특채"
            ],
            "title": "Result",
            "type": "string"
          }
        },
        "required": [
          "inspection_id",
          "result",
          "nonconformity_code",
          "lot_id",
          "lot_number",
          "ledger_entry_id"
        ],
        "title": "InspectionOut",
        "type": "object"
      },
...
      "TransportRefused": {
        "description": "400 · 404 · 405 · 500 의 본문 — 422 와 **같은 모양이고 이름 공간만 다르다**.",
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/TransportDetail"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "TransportRefused",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "읽는 엔드포인트가 서기 전까지 이 계약은 **좁혀질 수 있고 판은 움직이지 않는다.** 그 조각이 서는 날부터 이동 규칙이 선다 — 깨지면 앞자리, 호환되게 넓어지면 뒷자리.\n\n거절의 본문은 어느 경로에서나 `detail[]` 한 모양이다. `detail[].type` 이 기계가 읽는 자리이고 이름 공간이 셋이다 — 업무 규칙의 이름(경로마다 자기 열거를 든다 — 검사는 `Refusal`, 반품은 `ReturnRefusal`, 재검사는 `RetestRefusal`), 라우트 밖의 `Transport`, 그리고 pydantic 이 정한 이름(`missing` · `extra_forbidden` 등). 앞의 둘은 이 스펙이 열거로 들고, **그 열거에 없는 값은 셋째 무리**다.",
    "summary": "수입검사 한 건을 받아 판정하고, 합격이면 로트를 만든다. 공급사에 돌려보낸 것을 적고, 재고 로트였으면 원장에서 뺀다. 만료된 로트를 다시 검사해, 합격이면 새 만료일을 내고 불합격이면 잔량을 버린다.",
    "title": "조기경보 ERP — 관문 1 · 구매반품 · 재검사",
    "version": "0.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/inspections": {
      "post": {
        "description": "검사 한 건을 받는다.\n\n**받을 수 없는 것은 422 로 이름을 말하고 돌려보낸다.** 제약이 터지고 나서\n나오는 말은 제약 이름이라 검사원에게 아무것도 알려 주지 않는다 — 「왜 막혔는지\n모르는 실패」는 고칠 수 없는 실패다. **어떤 갈래를 무엇이 막는지는 한 자리에만\n적는다** — `app/services/incoming.py` 의 `RefusedInspection` 이 든다.",
        "operationId": "post_inspection_inspections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionOut"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRefused"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRefused"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRefused"
                }
              }
            },
            "description": "Method Not Allowed",
            "headers": {
              "Allow": {
                "description": "이 경로가 받는 메서드.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refused"
                }
              }
            },
            "description": "Unprocessable Entity",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRefused"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Post Inspection"
      }
    },
    "/purchase-returns": {
      "post": {
        "description": "공급사에 돌려보낸 것 한 건을 받는다.\n\n재고 로트였으면 반품 문서와 원장의 구매반품출고 줄이 함께 서고, 불합격분이면\n문서만 선다. **어떤 갈래를 무엇이 막는지는** `app/services/returns.py` 의\n`ReturnRefusal` 이 든다.",
        "operationId": "post_purchase_return_purchase_returns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseReturnIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseReturnOut"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRefused"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRefused"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRefused"
                }
              }
            },
            "description": "Method Not Allowed",
            "headers": {
              "Allow": {
                "description": "이 경로가 받는 메서드.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRefused"
                }
              }
            },
            "description": "Unprocessable Entity",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRefused"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "X-Request-Id": {
                "description": "이 요청을 가리키는 값. 보내면 그대로 돌아오고, 없거나 모양이 맞지 않으면 우리가 짓는다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Post Purchase Return"
      }
    },
...
  }
}